+++ biceps/src/main/java/org/somda/sdc/biceps/common/preprocessing/DescriptorChildRemover.java
@@ -4,7 +4,13 @@ import org.somda.sdc.biceps.common.MdibDescriptionModification;
 import org.somda.sdc.biceps.common.MdibDescriptionModifications;
 import org.somda.sdc.biceps.common.storage.DescriptionPreprocessingSegment;
 import org.somda.sdc.biceps.common.storage.MdibStorage;
-import org.somda.sdc.biceps.model.participant.*;
+import org.somda.sdc.biceps.model.participant.AbstractDescriptor;
+import org.somda.sdc.biceps.model.participant.AlertSystemDescriptor;
+import org.somda.sdc.biceps.model.participant.ChannelDescriptor;
+import org.somda.sdc.biceps.model.participant.MdsDescriptor;
+import org.somda.sdc.biceps.model.participant.ScoDescriptor;
+import org.somda.sdc.biceps.model.participant.SystemContextDescriptor;
+import org.somda.sdc.biceps.model.participant.VmdDescriptor;
 
 /**
  * Removes children from descriptors in order to avoid redundant information in the MDIB storage.
@@ -20,13 +26,13 @@ public class DescriptorChildRemover implements DescriptionPreprocessingSegment {
         } else if (descriptor instanceof VmdDescriptor) {
             removeChildren((VmdDescriptor) descriptor);
         } else if (descriptor instanceof ChannelDescriptor) {
-            removeChildren((ChannelDescriptor)descriptor);
+            removeChildren((ChannelDescriptor) descriptor);
         } else if (descriptor instanceof ScoDescriptor) {
-            removeChildren((ScoDescriptor)descriptor);
+            removeChildren((ScoDescriptor) descriptor);
         } else if (descriptor instanceof SystemContextDescriptor) {
-            removeChildren((SystemContextDescriptor)descriptor);
+            removeChildren((SystemContextDescriptor) descriptor);
         } else if (descriptor instanceof AlertSystemDescriptor) {
-            removeChildren((AlertSystemDescriptor)descriptor);
+            removeChildren((AlertSystemDescriptor) descriptor);
         }
     }
 
@@ -57,6 +63,14 @@ public class DescriptorChildRemover implements DescriptionPreprocessingSegment {
         descriptor.setAlertSystem(null);
     }
 
+    /**
+     * Removes the children from a provided mds.
+     * <p>
+     * Removes the battery, clock, system context, vmd, alert system and sco
+     * @param mds without the children
+     * @deprecated will be made private in 2.0.0
+     */
+    @Deprecated(since = "1.1.0", forRemoval = false)
     public void removeChildren(MdsDescriptor mds) {
         mds.setBattery(null);
         mds.setClock(null);