+++ biceps/src/main/java/org/somda/sdc/biceps/provider/preprocessing/CardinalityException.java
@@ -19,7 +19,10 @@ public class CardinalityException extends Exception {
         super(cause);
     }
 
-    public CardinalityException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+    public CardinalityException(
+            String message, Throwable cause,
+            boolean enableSuppression, boolean writableStackTrace
+    ) {
         super(message, cause, enableSuppression, writableStackTrace);
     }
 }