+++ biceps/src/main/java/org/somda/sdc/biceps/provider/preprocessing/HandleDuplicatedException.java
@@ -4,6 +4,7 @@ package org.somda.sdc.biceps.provider.preprocessing;
  * Exception that can be thrown if a handle is duplicated during preprocessing.
  */
 public class HandleDuplicatedException extends Exception {
+
     public HandleDuplicatedException() {
     }
 
@@ -19,7 +20,10 @@ public class HandleDuplicatedException extends Exception {
         super(cause);
     }
 
-    public HandleDuplicatedException(String message, Throwable cause, boolean enableSuppression, boolean writableStackTrace) {
+    public HandleDuplicatedException(
+            String message, Throwable cause,
+            boolean enableSuppression, boolean writableStackTrace
+    ) {
         super(message, cause, enableSuppression, writableStackTrace);
     }
 }